requirements.txt
absl-py==2.2.2
aiofiles==23.2.1
alembic==1.15.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
babel==2.17.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
blinker==1.7.0
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cssselect==1.2.0
cssselect2==0.8.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
executing==2.0.1
fastapi==0.110.2
feedfinder2==0.0.4
feedparser==6.0.11
ffmpy==0.3.2
filelock==3.13.4
Flask==3.0.3
flask-babel==4.0.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Moment==1.0.6
Flask-Reuploaded==1.4.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2024.3.1
gast==0.6.0
google-ai-generativelanguage==0.6.2
google-api-core==2.18.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
google-pasta==0.2.0
googleapis-common-protos==1.63.0
gradio==4.27.0
gradio_client==0.15.1
greenlet==3.0.3
grpcio==1.62.2
grpcio-status==1.62.2
gunicorn==23.0.0
h11==0.14.0
h5py==3.13.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.7
importlib_resources==6.4.0
imutils==0.5.4
ipykernel==6.29.4
ipython==8.23.0
itsdangerous==2.1.2
jedi==0.19.1
jieba3k==0.35.1
Jinja2==3.1.3
jiter==0.9.0
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
keras==3.9.2
kiwisolver==1.4.5
libclang==18.1.1
lxml==5.2.1
lxml_html_clean==0.1.1
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.4
openai==1.66.3
opencv-python-headless==4.9.0.80
opt_einsum==3.4.0
optree==0.15.0
orjson==3.10.1
packaging==24.0
-e git+https://github.com/eclipse/paho.mqtt.python@d45de3737879cfe7a6acc361631fa5cb1ef584bb#egg=paho_mqtt
pandas==2.2.2
parso==0.8.4
pika==1.3.2
pillow==11.1.0
platformdirs==4.2.0
prettytable==3.10.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
psycopg2==2.9.10
psycopg2-binary==2.9.10
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pydub==0.25.1
pydyf==0.11.0
Pygments==2.17.2
PyJWT==2.10.1
pyparsing==3.1.2
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-magic-bin==0.4.14
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.0
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
ruff==0.4.1
scikit-learn==1.6.1
scipy==1.13.0
seaborn==0.13.2
semantic-version==2.10.0
setuptools==78.1.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
stack-data==0.6.3
starlette==0.37.2
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.0.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tinysegmenter==0.3
tldextract==5.1.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
wcwidth==0.2.13
weasyprint==64.1
webencodings==0.5.1
websockets==11.0.3
Werkzeug==3.0.2
wheel==0.45.1
wrapt==1.17.2
WTForms==3.2.1
zopfli==0.2.3.post1
